Refactored locations to avoid duplication
